fix: correct type mismatch and add timestamps to projection

+++ src/main/java/br/com/stockshift/dto/warehouse/ProductWithStockResponse.java
@@ -30,7 +30,7 @@ public class ProductWithStockResponse {
     private Map<String, Object> attributes;
     private Boolean hasExpiration;
     private Boolean active;
-    private BigDecimal totalQuantity;
+    private Long totalQuantity;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
